+++ components/src/main/java/com/mercadolibre/android/andesui/bullet/AndesBullet.kt
@@ -2,7 +2,7 @@ package com.mercadolibre.android.andesui.bullet
 
 import android.text.SpannableStringBuilder
 
-class AndesBulletSpan(val startIndex: Int, val endIndex: Int) {
+class AndesBullet(val startIndex: Int, val endIndex: Int) {
     fun isValidRange(text: SpannableStringBuilder): Boolean {
         return (startIndex >= 0 &&
                 endIndex >= 0 &&