+++ demoapp/src/main/java/com/mercadolibre/android/andesui/demoapp/components/a11yplayground/ImportantForA11yActivity.kt
@@ -1,17 +1,17 @@
 package com.mercadolibre.android.andesui.demoapp.components.a11yplayground
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import android.widget.LinearLayout
 import com.mercadolibre.android.andesui.button.AndesButton
 import com.mercadolibre.android.andesui.demoapp.R
 
-class ImportantForA11yActivity : AppCompatActivity() {
+class ImportantForA11yActivity : A11yPlaygroundActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_important_for_a11y)
+        initActionBar()
         setupViews()
     }
 
@@ -24,5 +24,12 @@ class ImportantForA11yActivity : AppCompatActivity() {
             findViewById<AndesButton>(R.id.ifa_button_2).importantForAccessibility =
                 View.IMPORTANT_FOR_ACCESSIBILITY_NO_HIDE_DESCENDANTS
         }
+    }
+
+    private fun initActionBar() {
+        val navBarTitle = resources.getString(R.string.andes_demoapp_playground_title_important_for_a11y)
+        setSupportActionBar(findViewById(R.id.ifa_nav_bar))
+        supportActionBar?.title = navBarTitle
+        supportActionBar?.setDisplayHomeAsUpEnabled(true)
     }
 }