+++ demoapp/src/main/java/com/mercadolibre/android/andesui/demoapp/components/a11yplayground/SemanticViewsActivity.kt
@@ -1,15 +1,15 @@
 package com.mercadolibre.android.andesui.demoapp.components.a11yplayground
 
 import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.mercadolibre.android.andesui.demoapp.R
 
-class SemanticViewsActivity : AppCompatActivity() {
+class SemanticViewsActivity : A11yPlaygroundActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_semantic_views)
+        initActionBar()
         configFirstLayout()
         configSecondLayout()
     }
@@ -26,5 +26,12 @@ class SemanticViewsActivity : AppCompatActivity() {
             isFocusable = true
             contentDescription = "Top priced app. Four dollars and ninety-nine cents"
         }
+    }
+
+    private fun initActionBar() {
+        val navBarTitle = resources.getString(R.string.andes_demoapp_playground_title_semantic_views)
+        setSupportActionBar(findViewById(R.id.asv_nav_bar))
+        supportActionBar?.title = navBarTitle
+        supportActionBar?.setDisplayHomeAsUpEnabled(true)
     }
 }