+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # v3.35.0
+## 🛠 Fixes
+- Fix tag accessibility | Author: Andes Team
 ## 🚀 Feature
 - AndesBadgeIconPill hierarchy attribute added | Author: [@guidoperre-meli](https://github.com/guidoperre-meli)